resume_jd_analyser/requirements.txt
google-generativeai
pdfplumber
python-docx
numpy
streamlit
python-dotenv
